+++ Formula/curry.rb
@@ -6,9 +6,9 @@ class Curry < Formula
   # Update url + sha256 after tagging a release:
   #   git tag v1.2.6 && git push origin v1.2.6
   #   curl -L https://github.com/deconstructo/curry/archive/refs/tags/v1.6.2.tar.gz | shasum -a 256
-  url "https://github.com/deconstructo/curry/archive/refs/tags/v1.10.2.tar.gz"
-  sha256 "e61bc9f216b97ad8712e490f13b46b1a40e8b75295e1336565b5a255678ab594"
-  version "1.10.2"
+  url "https://github.com/deconstructo/curry/archive/refs/tags/v1.10.3.tar.gz"
+  sha256 "0000000000000000000000000000000000000000000000000000000000000"
+  version "1.10.3"
 
   head "https://github.com/deconstructo/curry.git", branch: "main"